backend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
databento>=0.45.0
python-dotenv>=1.0.0
